requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2
django-bootstrap4==21.1
djangorestframework==3.12.4
Pillow==8.4.0
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
